attic: Rename stream to io
