+++ attic/Cargo.toml
@@ -53,13 +53,13 @@ nix-main = { version = "2.24", feature = "nix_store" }
 [features]
 default = [
 	"chunking",
+	"io",
 	"nix_store",
-	"stream",
 	"tokio",
 ]
 
 # Chunking.
-chunking = ["tokio", "stream", "dep:async-stream"]
+chunking = ["tokio", "io", "dep:async-stream"]
 
 # Native libnixstore bindings.
 #
@@ -73,8 +73,8 @@ nix_store = [
 	"dep:system-deps",
 ]
 
-# Stream utilities.
-stream = ["tokio", "dep:async-stream"]
+# IO utilities.
+io = ["tokio", "dep:async-stream"]
 
 # Tokio runtime.
 tokio = ["dep:tokio", "tokio/rt", "tokio/time"]
